Added a small icon for `Templates` directory.
In Thunar, the directories in `$HOME` have special icons different from the default directory icon. These are of size 128×128, and are all teal in colour. Likewise, the same directories, when seen from an 'Open File' dialog, have special icons, but these are of size 16×16, and are grey in colour. Except for the `Templates` directory, which uses the teal 128×128 icon instead! To fix this, I created a symbolic link to the correct 16×16 grey icon. However, that did not work! Out of curiosity, I changed `MinSize` of `[128x128/places]` (in `Teal-Sea/Teal-Sea-Light-Icons/index.theme`) from 8 to 24. Suddenly, the 'Open File' dialog started using the correct (grey) icon for `Templates`. When I changed 24 back to 8, nothing broke: the icon is still grey, so I think there was some theming issue, and just creating the symbolic link was enough.